@@ -2,7 +2,7 @@ name: network_info_plus_example
 description: Demonstrates how to use the network_info_plus plugin.
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.18.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -16,7 +16,7 @@ dev_dependencies:
     sdk: flutter
   integration_test:
     sdk: flutter
-  test: ^1.16.4
+  test: ^1.22.0
   flutter_lints: ^2.0.1
 
 flutter:
