@@ -6,8 +6,8 @@ repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
 issue_tracker: https://github.com/fluttercommunity/plus_plugins/labels/network_info_plus
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.11.0"
+  sdk: ">=2.18.0 <4.0.0"
+  flutter: ">=3.3.0"
 
 flutter:
   plugin:
@@ -28,7 +28,7 @@ flutter:
         fileName: src/network_info_plus_web.dart
 
 dependencies:
-  collection: ^1.15.0
+  collection: ^1.17.0
   nm: ^0.5.0
   flutter:
     sdk: flutter
@@ -36,9 +36,10 @@ dependencies:
     sdk: flutter
   meta: ^1.8.0
   network_info_plus_platform_interface: ^1.1.3
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^5.3.2
-  plugin_platform_interface: ^2.1.3
+  mockito: ^5.4.0
+  plugin_platform_interface: ^2.1.4
   flutter_lints: ^2.0.1
